split app into app and server
