+++ app.js
@@ -1,56 +1,51 @@
-require('dotenv').config()
-process.env["NODE_TLS_REJECT_UNAUTHORIZED"] = 0;
-const express = require('express');
-const app = express();
-const mongoose = require('mongoose');
-const methodOverride = require('method-override');
-const https = require('https');
-const fs = require('fs');
-const key = fs.readFileSync('key.pem');
-const cert = fs.readFileSync('cert.pem');
-const server = https.createServer({key: key, cert: cert }, app);
+module.exports = (database) =>{
+    const express = require('express');
+    const app = express();
+    const methodOverride = require('method-override');
+    const mongoose = require('mongoose');
 
-mongoose.connect(process.env.DATABASE_URL, { useNewUrlParser: true })
-const db = mongoose.connection
-db.on('error', (error) => console.error(error))
-db.once('open', () => console.log('Connected to Database'))
-const cookieParser = require('cookie-parser');
-app.use(cookieParser());
-app.use(methodOverride('X-HTTP-Method-Override'))
-app.use(express.json())
-app.use(express.urlencoded({extended : true}))
-app.use(methodOverride('_method'))
+    mongoose.connect(database, { useNewUrlParser: true })
+    const db = mongoose.connection
+    db.on('error', (error) => console.error(error))
+    db.once('open', () => console.log('Connected to Database'))
 
-const usersRouter = require('./routes/users').router
-app.use('/users', usersRouter)
+    app.use(methodOverride('X-HTTP-Method-Override'))
+    app.use(express.json())
+    app.use(express.urlencoded({extended : true}))
+    app.use(methodOverride('_method'))
 
-const recordsRouter = require('./routes/records')
-app.use('/records', recordsRouter)
-app.set('view engine', 'ejs');
+    const usersRouter = require('./routes/users').router
+    app.use('/users', usersRouter)
 
-path = require('path')
+    const recordsRouter = require('./routes/records')
+    app.use('/records', recordsRouter)
 
-app.use('/css', express.static(path.join(__dirname, '/css')));
-app.use("/static", express.static('./static/'));
+    path = require('path')
+    app.use('/css', express.static(path.join(__dirname, '/css')));
+    app.use("/static", express.static('./static/'));
 
-app.get("/register", (req, res) =>{
-    res.render("register.ejs")
-})
-app.get("/", (req, res) => {
-    //res.render("index.ejs", {data : req.data})
-    res.redirect("/home")
-})
-app.get("/login", (req, res) =>{
-    res.render("login.ejs")
-  })
-app.get("/home", (req, res) => {
-    res.render("home.ejs")
-})
-app.get("/lol", (req, res) => {
-    res.render("index.ejs", {data : req.data})
-})
+    app.set('view engine', 'ejs');
 
-app.get("/users/details", async(req, res) => {
-    res.render("detailsUser.ejs")
-})
-server.listen(3000, () => { console.log('listening on 3000') });
+    app.get("/register", (req, res) =>{
+        res.render("register.ejs")
+    })
+    app.get("/", (req, res) => {
+        res.redirect("/home")
+    })
+    app.get("/login", (req, res) =>{ 
+        res.render("login.ejs")
+    })
+    app.get("/home", (req, res) => {
+        res.render("home.ejs")
+    })
+    app.get("/lol", (req, res) => {
+        res.render("index.ejs", {data : req.data})
+    })
+
+    app.get("/users/details", async(req, res) => {
+        res.render("detailsUser.ejs")
+    })
+    return app
+};
+
+
